Renamed validation.js to validators.js as requested in PR
